refactor: move Node classes to _model.py
