+++ src/molgenis/bbmri_eric/nodes.py
@@ -1,73 +1,6 @@
-from dataclasses import dataclass
 from typing import List
 
-
-@dataclass(frozen=True)
-class Node:
-    code: str
-    package = "eu_bbmri_eric"
-
-    @property
-    def persons_staging_id(self) -> str:
-        return self._get_staging_id("persons")
-
-    @property
-    def networks_staging_id(self) -> str:
-        return self._get_staging_id("networks")
-
-    @property
-    def biobanks_staging_id(self) -> str:
-        return self._get_staging_id("biobanks")
-
-    @property
-    def collections_staging_id(self) -> str:
-        return self._get_staging_id("collections")
-
-    def _get_staging_id(self, simple_name: str) -> str:
-        return f"eu_bbmri_eric_{self.code}_{simple_name}"
-
-    def get_staging_table_ids(self) -> List[str]:
-        return [
-            self.persons_staging_id,
-            self.networks_staging_id,
-            self.biobanks_staging_id,
-            self.collections_staging_id,
-        ]
-
-
-@dataclass(frozen=True)
-class ExternalNode(Node):
-    # TODO rename to IndependentNode
-    url: str
-
-    @property
-    def persons_external_id(self) -> str:
-        return self._get_id("persons")
-
-    @property
-    def networks_external_id(self) -> str:
-        return self._get_id("networks")
-
-    @property
-    def biobanks_external_id(self) -> str:
-        return self._get_id("biobanks")
-
-    @property
-    def collections_external_id(self) -> str:
-        return self._get_id("collections")
-
-    @staticmethod
-    def _get_id(simple_name: str) -> str:
-        return f"eu_bbmri_eric_{simple_name}"
-
-    def get_external_table_ids(self) -> List[str]:
-        return [
-            self.persons_external_id,
-            self.networks_external_id,
-            self.biobanks_external_id,
-            self.collections_external_id,
-        ]
-
+from molgenis.bbmri_eric._model import ExternalNode, Node
 
 _external_nodes = {
     "BE": ExternalNode("BE", "https://directory.bbmri.be"),
